refactor JumpstartAddAssets, remove JumpstartIntro component
