+++ src/jumpstart/JumpstartAddAssets.tsx
@@ -1,29 +1,52 @@
-import React, { useRef } from 'react'
+import React, { useEffect, useRef } from 'react'
 import { useTranslation } from 'react-i18next'
+import { ScrollView, StyleSheet, Text, View } from 'react-native'
+import { SafeAreaView, useSafeAreaInsets } from 'react-native-safe-area-context'
+import { useDispatch } from 'react-redux'
 import AppAnalytics from 'src/analytics/AppAnalytics'
 import { JumpstartEvents } from 'src/analytics/Events'
 import AddAssetsBottomSheet, { AddAssetsAction } from 'src/components/AddAssetsBottomSheet'
+import BackButton from 'src/components/BackButton'
 import { BottomSheetRefType } from 'src/components/BottomSheet'
 import Button, { BtnSizes, BtnTypes } from 'src/components/Button'
+import CustomHeader, { CUSTOM_HEADER_HEIGHT } from 'src/components/header/CustomHeader'
 import { CICOFlow, FiatExchangeFlow } from 'src/fiatExchanges/utils'
-import JumpstartIntro from 'src/jumpstart/JumpstartIntro'
+import Leaf from 'src/images/Leaf'
+import WaveCurve from 'src/images/WaveCurve'
+import { jumpstartIntroSeen } from 'src/jumpstart/slice'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import { isAppSwapsEnabledSelector } from 'src/navigator/selectors'
 import { useSelector } from 'src/redux/hooks'
+import Colors from 'src/styles/colors'
+import { typeScale } from 'src/styles/fonts'
+import { Spacing } from 'src/styles/styles'
+import { jumpstartSendTokensSelector } from 'src/tokens/selectors'
 import { TokenActionName } from 'src/tokens/types'
 import networkConfig from 'src/web3/networkConfig'
 
 export default function JumpstartAddAssets() {
+  const { t } = useTranslation()
+  const dispatch = useDispatch()
+  const insets = useSafeAreaInsets()
   const addAssetsBottomSheetRef = useRef<BottomSheetRefType>(null)
 
-  const { t } = useTranslation()
-
+  const tokens = useSelector(jumpstartSendTokensSelector)
   const isSwapEnabled = useSelector(isAppSwapsEnabledSelector)
+  const noTokens = tokens.length === 0
+
+  // Track in analytics whenever user sees the intro
+  useEffect(() => {
+    AppAnalytics.track(JumpstartEvents.jumpstart_intro_seen)
+  }, [])
 
   const handleShowAddFunds = () => {
     AppAnalytics.track(JumpstartEvents.jumpstart_add_assets_show_actions)
     addAssetsBottomSheetRef.current?.snapToIndex(0)
+  }
+
+  const onIntroDismiss = () => {
+    dispatch(jumpstartIntroSeen())
   }
 
   const addAssetsActions: AddAssetsAction[] = [
@@ -71,25 +94,81 @@ export default function JumpstartAddAssets() {
   ]
 
   return (
-    <JumpstartIntro
-      showNoAssetsHint
-      button={
+    <SafeAreaView style={styles.safeAreaContainer} edges={['top']}>
+      <Leaf style={styles.palmImage} />
+      <WaveCurve style={styles.waveImage} />
+      <CustomHeader style={styles.header} left={<BackButton />} />
+
+      <ScrollView contentContainerStyle={[styles.container, { paddingBottom: insets.bottom }]}>
+        <Text style={styles.title}>{t('jumpstartIntro.title')}</Text>
+
+        <View style={styles.description}>
+          <Text style={styles.descriptionLine}>{t('jumpstartIntro.description')}</Text>
+          {noTokens && (
+            <Text style={styles.descriptionLine}>{t('jumpstartIntro.noFundsHint')}</Text>
+          )}
+        </View>
+
         <Button
-          testID="JumpstartAddAssets/noFundsButton"
-          onPress={handleShowAddFunds}
-          text={t('jumpstartIntro.addFundsCelo.cta')}
+          testID={
+            noTokens ? 'JumpstartAddAssets/noFundsButton' : 'JumpstartEnterAmount/haveFundsButton'
+          }
+          text={
+            noTokens ? t('jumpstartIntro.addFundsCelo.cta') : t('jumpstartIntro.haveFundsButton')
+          }
+          onPress={noTokens ? handleShowAddFunds : onIntroDismiss}
           type={BtnTypes.PRIMARY}
           size={BtnSizes.FULL}
         />
-      }
-    >
-      <AddAssetsBottomSheet
-        forwardedRef={addAssetsBottomSheetRef}
-        title={t('jumpstartIntro.addFundsCelo.title')}
-        description={t('jumpstartIntro.addFundsCelo.description')}
-        actions={addAssetsActions}
-        testId="Jumpstart/addFundsCeloBottomSheet"
-      />
-    </JumpstartIntro>
+      </ScrollView>
+
+      {noTokens && (
+        <AddAssetsBottomSheet
+          forwardedRef={addAssetsBottomSheetRef}
+          title={t('jumpstartIntro.addFundsCelo.title')}
+          description={t('jumpstartIntro.addFundsCelo.description')}
+          actions={addAssetsActions}
+          testId="Jumpstart/addFundsCeloBottomSheet"
+        />
+      )}
+    </SafeAreaView>
   )
 }
+
+const styles = StyleSheet.create({
+  safeAreaContainer: {
+    flex: 1,
+  },
+  header: {
+    paddingHorizontal: Spacing.Thick24,
+  },
+  container: {
+    paddingHorizontal: Spacing.Thick24,
+    flex: 1,
+    gap: Spacing.Regular16,
+    justifyContent: 'center',
+    marginTop: -CUSTOM_HEADER_HEIGHT,
+  },
+  title: {
+    ...typeScale.titleMedium,
+    color: Colors.black,
+    textAlign: 'center',
+  },
+  description: {
+    gap: 4,
+  },
+  descriptionLine: {
+    ...typeScale.bodyMedium,
+    textAlign: 'center',
+  },
+  palmImage: {
+    position: 'absolute',
+    top: 0,
+    right: 20,
+  },
+  waveImage: {
+    position: 'absolute',
+    bottom: 0,
+    left: 0,
+  },
+})
